internal → public, Deterrariation
